# salesman-ecommerce/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product/detail/{product_id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "View detail product by id",
                "operationId": "1a276264aadfc387a42a8111572a866c",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "product_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/product/list/owner/{owner_id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "View product by owner",
                "operationId": "b6eff1db2c568a2c5447f4d5353dccbd",
                "parameters": [
                    {
                        "name": "owner_id",
                        "in": "path",
                        "description": "owner_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/product/create": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create product",
                "operationId": "efe61ac114f4445ae99b719abc8ca39d",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "T-shirt"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "owner_id": {
                                                "type": "string",
                                                "example": "USER00000003"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "category": {
                                                "type": "string",
                                                "example": "CATE001"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "size": {
                                                "type": "string",
                                                "example": "M"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "price": {
                                                "type": "float",
                                                "example": "53.2"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "description": {
                                                "type": "text",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "image": {
                                                "description": "",
                                                "type": "string",
                                                "format": "binary"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/product/update/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "operationId": "779d39a2b74c12cc494f0bff640cf45e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "T-shirt"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "owner_id": {
                                                "type": "string",
                                                "example": "USER00000003"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "category": {
                                                "type": "string",
                                                "example": "CATE001"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "size": {
                                                "type": "string",
                                                "example": "M"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "price": {
                                                "type": "float",
                                                "example": "53.2"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "description": {
                                                "type": "text",
                                                "example": ""
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "image": {
                                                "description": "",
                                                "type": "string",
                                                "format": "binary"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/product/delete/{product_id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product",
                "operationId": "13a3f431f60ded1d070ea6815f1acecc",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}